feat: fix cypress fixtures, rename file
