+++ cypress/e2e/404.cy.ts
@@ -1,4 +1,4 @@
-import { configureSkylarkIntercepts } from "../support/utils/handlers";
+import { configureSkylarkIntercepts } from "../support/utils/intercepts";
 
 describe("404 page", () => {
   beforeEach(() => {
