+++ cypress/e2e/contentLibrary/createObjectModal.cy.ts
@@ -1,4 +1,4 @@
-import { configureSkylarkIntercepts } from "../../support/utils/handlers";
+import { configureSkylarkIntercepts } from "../../support/utils/intercepts";
 
 describe("Create Object Modal", () => {
   beforeEach(() => {
