+++ cypress/e2e/contentLibrary/dragdrop.cy.ts
@@ -2,7 +2,7 @@ import {
   hasMatchingVariable,
   hasOperationName,
 } from "../../support/utils/graphqlTestUtils";
-import { configureSkylarkIntercepts } from "../../support/utils/handlers";
+import { configureSkylarkIntercepts } from "../../support/utils/intercepts";
 
 const allDevicesAllCustomersAvailability =
   "Always - All devices, all non-kids customer types";
