+++ cypress/e2e/contentLibrary/panel/tabs/availability.cy.ts
@@ -1,7 +1,7 @@
 import {
   allDevicesAllCustomersAvailability,
   configureSkylarkIntercepts,
-} from "../../../../support/utils/handlers";
+} from "../../../../support/utils/intercepts";
 
 describe("Content Library - Object Panel - Availability tab", () => {
   beforeEach(() => {
