+++ cypress/e2e/contentLibrary/panel/tabs/availabilityDimensions.cy.ts
@@ -2,7 +2,7 @@ import { hasOperationName } from "../../../../support/utils/graphqlTestUtils";
 import {
   allDevicesAllCustomersAvailability,
   configureSkylarkIntercepts,
-} from "../../../../support/utils/handlers";
+} from "../../../../support/utils/intercepts";
 
 describe("Content Library - Object Panel - Availability Dimensions tab", () => {
   beforeEach(() => {
