+++ cypress/e2e/contentLibrary/panel/tabs/metadata.cy.ts
@@ -1,5 +1,5 @@
 import { hasOperationName } from "../../../../support/utils/graphqlTestUtils";
-import { configureSkylarkIntercepts } from "../../../../support/utils/handlers";
+import { configureSkylarkIntercepts } from "../../../../support/utils/intercepts";
 
 describe("Content Library - Object Panel - Metadata tab", () => {
   beforeEach(() => {
