+++ cypress/e2e/contentLibrary/panel/tabs/playback.cy.ts
@@ -1,4 +1,4 @@
-import { configureSkylarkIntercepts } from "../../../../support/utils/handlers";
+import { configureSkylarkIntercepts } from "../../../../support/utils/intercepts";
 
 describe("Content Library - Object Panel - Playback tab", () => {
   beforeEach(() => {
