+++ cypress/e2e/contentLibrary/search.cy.ts
@@ -3,7 +3,7 @@ import {
   hasMatchingVariable,
   hasOperationName,
 } from "../../support/utils/graphqlTestUtils";
-import { configureSkylarkIntercepts } from "../../support/utils/handlers";
+import { configureSkylarkIntercepts } from "../../support/utils/intercepts";
 
 const assetOnlyQuery =
   "query SEARCH($ignoreAvailability: Boolean = true, $queryString: String!) {\n  search(\n    ignore_availability: $ignoreAvailability\n    query: $queryString\n    limit: 1000\n  ) {\n    __typename\n    objects {\n      ... on Asset {\n        __typename\n        uid\n        external_id\n        __Asset__slug: slug\n        __Asset__title: title\n        __Asset__type: type\n        __Asset__url: url\n      }\n      __typename\n    }\n  }\n}";
@@ -16,7 +16,7 @@ describe("Content Library - Search", () => {
 
     cy.visit("/");
     cy.wait("@introspectionQuery");
-    cy.wait("@getObjectConfig");
+    cy.wait("@getObjectsConfig");
 
     cy.get("[data-testid=animated-skylark-logo]").should("not.exist");
   });
