+++ cypress/e2e/navigation.cy.ts
@@ -1,4 +1,4 @@
-import { configureSkylarkIntercepts } from "../support/utils/handlers";
+import { configureSkylarkIntercepts } from "../support/utils/intercepts";
 
 describe("Navigation", () => {
   beforeEach(() => {
@@ -12,12 +12,12 @@ describe("Navigation", () => {
 
   it("contains the navigation bar", () => {
     cy.contains("Skylark");
-    cy.get("nav").get("ul").find("li").should("have.length", 2);
+    cy.get("nav").find("li").should("have.length", 2);
   });
 
   it("opens the navigation bar on mobile", () => {
     cy.viewport("iphone-xr");
     cy.get("#mobile-nav-toggle").click();
-    cy.get("nav").get("ul").find("li").should("have.length", 2);
+    cy.get("nav").find("li").should("have.length", 2);
   });
 });
